feature: enable limited ads
With GDPR a user can choose to not give permission to have his private
data sent to tracking providers. For this purpose Google has implemented
the limited ads privacy setting. Choosing this setting comes with a
changed URL for the GPT source, ad additional "ltd" parameter for all
xhr calls to the actual ads and will prevent sending cookies to the ads
server.

Signed-off-by: Tobias Winkler <tobias.winkler@omio.com>